# common/openapi-api.yaml
openapi: 3.0.3
info:
  version: v1
  title: Solarpunk Drifters API
servers:
  - url: http://localhost:3000/api
    description: Local dev server (with local reverse proxy)
  - url: https://solarpunkdrifters.bond/api
    description: Production server (uses live data)
paths:
...
  /expeditions/encounter-card-choice:
    post:
      description: Make a choice on the active encounter card
      requestBody:
        description: A player choice for the active encounter card
        content:
          application/json:
            schema:
              type: object
              properties:
                choice:
                  type: integer
                  minimum: 0
      responses:
        '200':
          description: Successfully executed a choice.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpeditionUpdate'
  /expeditions/play-card:
    post:
      description: Play a Drifter Card in the player's hand (between encounters)
      requestBody:
        description: A request to play a Drifter Card
        content:
          application/json:
            schema:
              type: object
              properties:
                cardId:
                  type: string
      responses:
        '200':
          description: Successfully played a Drifter Card.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpeditionUpdate'
...
security:
  - sessionCookieAuth: []
components:
  securitySchemes:
    sessionCookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
  schemas:
...
    ClientEvent:
      description: Describes an event that happened in the game, relevant to the client
      oneOf:
        - $ref: '#/components/schemas/ClientEventRollResult'
    ClientEventRollResult:
      type: object
      properties:
        type:
          type: string
          enum: ['CLIENT_EVENT_ROLL_RESULT']
        payload:
          $ref: '#/components/schemas/RollResult'
      required: [type, payload]
...
    ExpeditionUpdate:
      description: >-
        A list of client events, and a game "update" in JSON Patch (RFC 6902) format,
        representing a sequence of operations to apply to the GameState
      type: object
      properties:
        clientEvents:
          description: A list of game events relevant to the client
          type: array
          items:
            $ref: '#/components/schemas/ClientEvent'
        update:
          $ref: '#/components/schemas/PatchRequest'

    # GameMode:
    #   type: string
    #   enum: [ACTIVE_ENCOUNTER, BETWEEN_ENCOUNTERS, LOADOUT]
...
    RollResult:
      type: object
      properties:
        rolls:
          type: array
          items:
            type: integer
        outcome:
          $ref: '#/components/schemas/RollOutcome'
      required: [rolls, outcome]
    RollOutcome:
      type: string
      enum:
        - 'ROLL_OUTCOME_FAILURE'
        - 'ROLL_OUTCOME_MIXED_SUCCESS'
        - 'ROLL_OUTCOME_STRONG_SUCCESS'
...
    PatchRequest:
      type: array
      items:
        oneOf:
          - $ref: '#/components/schemas/JSONPatchRequestAddReplaceTest'
          - $ref: '#/components/schemas/JSONPatchRequestRemove'
          - $ref: '#/components/schemas/JSONPatchRequestMoveCopy'
    JSONPatchRequestAddReplaceTest:
      type: object
      additionalProperties: false
      required:
        - value
        - op
        - path
      properties:
        path:
          description: A JSON Pointer path.
          type: string
        value:
          description: The value to add, replace or test.
        op:
          description: The operation to perform.
          type: string
          enum:
            - add
            - replace
            - test
    JSONPatchRequestRemove:
      type: object
      additionalProperties: false
      required:
        - op
        - path
      properties:
        path:
          description: A JSON Pointer path.
          type: string
        op:
          description: The operation to perform.
          type: string
          enum:
            - remove
    JSONPatchRequestMoveCopy:
      type: object
      additionalProperties: false
      required:
        - from
        - op
        - path
      properties:
        path:
          description: A JSON Pointer path.
          type: string
        op:
          description: The operation to perform.
          type: string
          enum:
            - move
            - copy
        from:
          description: A JSON Pointer path.
          type: string